backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
httpx>=0.28.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyjwt>=2.8.0
pyotp>=2.9.0
qrcode>=7.4.0
bcrypt>=4.0.0
# File parsing
openpyxl>=3.1.0
python-docx>=1.1.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
# Email
aiosmtplib>=3.0.0
email-validator>=2.0.0
# Database (PostgreSQL support)
psycopg2-binary>=2.9.0
# Error tracking
sentry-sdk[fastapi]>=1.0.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
